backend/pyproject.toml
[project]
name = "kratzbaum"
version = "0.1.0"
description = "Self-hosted plant management system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "apscheduler>=3.10.0",
    "pywebpush>=2.0.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "pydantic-settings>=2.6.0",
    "alembic>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
